EDSC-4475: Converting AboutCwicModal
